+++ apps/web/app/composables/useCancellationForm/useCancellationForm.ts
@@ -1,16 +1,45 @@
 import type { ApiError } from '@plentymarkets/shop-api';
-import type { SubmitCancellation, UseCancellationFormReturn, UseCancellationFormState } from './types';
+import { userGetters } from '@plentymarkets/shop-api';
+import { toTypedSchema } from '@vee-validate/yup';
+import { object, string } from 'yup';
+import type { SubmitCancellation, UseCancellationFormState } from './types';
 
-export const useCancellationForm: UseCancellationFormReturn = () => {
+export const useCancellationForm = () => {
   const state = useState<UseCancellationFormState>('useCancellationForm', () => ({
     loading: false,
   }));
 
+  const { getSetting } = useSiteSettings('cloudflareTurnstileApiSiteKey');
+  const turnstileSiteKey = getSetting() ?? '';
+
+  const validationSchema = toTypedSchema(
+    object({
+      orderId: string()
+        .trim()
+        .required(t('error.requiredField'))
+        .matches(/^[1-9][0-9]*$/, t('cancellationForm.orderIdInvalid'))
+        .default(''),
+      name: string().trim().required(t('error.requiredField')).default(''),
+      email: string()
+        .trim()
+        .required(t('error.email.required'))
+        .test('is-valid-email', t('storefrontError.contactMail.emailInvalid'), (mail: string) =>
+          userGetters.isValidEmailAddress(mail),
+        )
+        .default(''),
+      turnstile:
+        turnstileSiteKey.length > 0
+          ? string().required(t('error.turnstileRequired')).default('')
+          : string().optional().default(''),
+    }),
+  );
+
   const submitCancellation: SubmitCancellation = async (params) => {
     state.value.loading = true;
     try {
-      const { data } = await useSdk().plentysystems.doSubmitCancellation(params);
-      return data?.email ?? null;
+        //const { data } = await useSdk().plentysystems.doSubmitCancellation(params);
+        const data = { email: '' }; // TODO: replace with actual API call once backend is implemented
+        return data?.email ?? null;
     } catch (error) {
       useHandleError(error as ApiError);
       return null;
@@ -21,6 +50,12 @@ export const useCancellationForm: UseCancellationFormReturn = () => {
 
   return {
     submitCancellation,
+    validationSchema,
+    turnstileSiteKey,
     ...toRefs(state.value),
   };
 };
+
+function sleep(arg0: number) {
+    throw new Error('Function not implemented.');
+}
